refactor: types 폴더 구조 개선 및 파일명 카멜 케이스로 통일
