+++ frontend/src/components/@common/buttons/copyButton/CopyButton.styles.ts
@@ -1,7 +1,7 @@
 import type { Theme } from '@emotion/react';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import type { CopyButtonVariant } from '../../../../types/buttonTypes';
+import type { CopyButtonVariant } from '../../../../types/button.type';
 
 const color = (theme: Theme) => ({
   filled: css`
