+++ frontend/src/types/globals/client.d.ts
