+++ frontend/src/types/spaceContent.type.ts
